requirements.txt
streamlit>=1.33
streamlit-webrtc
mediapipe
openai-whisper
opencv-python
pandas
matplotlib
pytest
